+++ internal/cli/cmd/lsp.go
@@ -13,12 +13,13 @@ type LSP struct {
 
 func (cmd *LSP) Run(container *cli.Container) error {
 	server := lsp.NewServer(lsp.ServerOpts{
-		Name:      "zk",
-		Version:   container.Version,
-		Logger:    container.Logger,
-		LogFile:   opt.NewNotEmptyString(cmd.Log),
-		Notebooks: container.Notebooks,
-		FS:        container.FS,
+		Name:           "zk",
+		Version:        container.Version,
+		Logger:         container.Logger,
+		LogFile:        opt.NewNotEmptyString(cmd.Log),
+		Notebooks:      container.Notebooks,
+		TemplateLoader: container.TemplateLoader,
+		FS:             container.FS,
 	})
 
 	return server.Run()